:recycle:Refactor: 게시글 수 응답하도록 마이페이지 api 수정

+++ src/main/java/com/practice/likelionhackathoncesco/domain/user/service/UserCreditService.java
@@ -25,7 +25,7 @@ public class UserCreditService {
         userRepository
             .findById(userId)
             .orElseThrow(() -> new CustomException(UserErrorCode.USER_NOT_FOUND));
-    Long postCount = user.getPostCount();
+    Integer postCount = user.getPostCount();
 
     log.info("[UserCreditService] 사용자별 생성한 게시글 개수 조회 : postCount={}", postCount);
